examples/docker_submission/docker_deepgaze3/requirements.txt
cython
flask
gunicorn
numpy

# Add additional dependencies here
pysaliency
scipy
torch
flask_orjson
git+https://github.com/matthias-k/deepgaze